auth/pageをSSRとする設計に変更
